requirements.txt
albumentations==1.3.1
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
idna==3.4
imageio==2.31.6
importlib-resources==6.1.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.2
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==1.13.0+cu117
torchaudio==0.13.0+cu117
torchvision==0.14.0+cu117
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zipp==3.17.0
